more light refactors
