+++ debate/src-js/SpanSelection2.scala
@@ -1,5 +1,7 @@
 package debate
 
+import annotation.unused
+
 import jjm.ling.ISpan
 
 import japgolly.scalajs.react.vdom.html_<^._
@@ -21,9 +23,6 @@ object SpanSelection2 {
     // in order for the lenses to actually compile. something is messed up
     // private[this] val indexSet: (Index => Selecting => Selecting) =
     //   (i: Index) => (s: Selecting) => s.copy(index = i)
-    (a: Int) => (s: Selecting) => s.copy(anchor = a)
-    (e: Int) => (s: Selecting) => s.copy(endpoint = e)
-    // val index = Lens[Selecting, Index](_.index)(i => s => s.copy(index = i))
     val anchor = Lens[Selecting, Int](_.anchor)(a => s => s.copy(anchor = a))
     val endpoint =
       Lens[Selecting, Int](_.endpoint)(e => s => s.copy(endpoint = e))
@@ -60,7 +59,7 @@ object SpanSelection2 {
 
   class Backend(scope: BackendScope[Props, Status]) {
 
-    def hover(props: Props, state: Status)(endpoint: Int) =
+    def hover(@unused props: Props, @unused state: Status)(endpoint: Int) =
       scope.modState {
         case Selecting(anchor, _) => Selecting(anchor, endpoint)
         case x                    => x
